feat: set up webpack for jest
